Move hooks to appropriate directories
